refactor: rename hook files
