+++ src/features/mcp-servers/hooks/use-mutation-restart-server.ts
