+++ src/features/mcp-servers/hooks/use-mutation-stop-server.ts
